stage1/kvm/qemu: Add patch file for 9pfs getattr/unlink/fattr fix
Doing an fstat on an unlinked file can fail on 9pfs, as the
lookup is done with the file path, not the file handle. If the
file handle is available, use it.
This patch from Eric Van Hensbergen for lkvm comes from:
 https://github.com/ericvh/linux-kvm/commit/2fa2f7e728ac08a7d9006516870db1a986aa6acc

And then looks like it was ported to QEMU by Greg Kurz?