tests: rkt_stop: logging and timing for semaphore kvm diagnosis
This test hangs up on semaphoreCI often under kvm. In order to try
and track down more details, add some timing information, and if
we timeout, wait a bit more.
Also, increase the granularity of the first timing loop so we can
get a feel for how long it actually takes, normally....

Signed-off-by: Graham Whaley <graham.whaley@intel.com>